При переключении города перемещает карту
