+++ src/view/components/map/const.ts
@@ -1,4 +1,24 @@
+import { Icon } from 'leaflet';
+
 const URL_MARKER_DEFAULT = '../../../../public/img/pin.svg';
 const URL_MARKER_ACTIVE = '../../../../public/img/pin-active.svg';
 
-export { URL_MARKER_ACTIVE, URL_MARKER_DEFAULT};
+const offerMapStyle = {
+  width: '1144px',
+  marginLeft: 'auto',
+  marginRight: 'auto',
+};
+
+const defaultCustomIcon = new Icon({
+  iconUrl: URL_MARKER_DEFAULT,
+  iconSize: [30, 40],
+  iconAnchor: [20, 40]
+});
+
+const currentCustomIcon = new Icon({
+  iconUrl: URL_MARKER_ACTIVE,
+  iconSize: [30, 40],
+  iconAnchor: [20, 40]
+});
+
+export { defaultCustomIcon, currentCustomIcon, offerMapStyle};
