+++ src/view/components/map/map.tsx
@@ -1,10 +1,10 @@
-import { Icon, layerGroup, Marker } from 'leaflet';
+import { layerGroup, Marker } from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 import { CityType, OfferType } from '../../../types/types.ts';
 import { useEffect, useRef } from 'react';
 import useMap from './use-map.ts';
 import { NameMap } from '../../../const.ts';
-import { URL_MARKER_ACTIVE, URL_MARKER_DEFAULT } from './const.ts';
+import { currentCustomIcon, defaultCustomIcon, offerMapStyle } from './const.ts';
 
 type MapProps = {
   city: CityType;
@@ -13,27 +13,19 @@ type MapProps = {
   mapName: keyof typeof NameMap;
 }
 
-const offerMapStyle = {
-  width: '1144px',
-  marginLeft: 'auto',
-  marginRight: 'auto',
-};
-
-const defaultCustomIcon = new Icon({
-  iconUrl: URL_MARKER_DEFAULT,
-  iconSize: [30, 40],
-  iconAnchor: [20, 40]
-});
-
-const currentCustomIcon = new Icon({
-  iconUrl: URL_MARKER_ACTIVE,
-  iconSize: [30, 40],
-  iconAnchor: [20, 40]
-});
-
 function Map({city, offers, activeOffer, mapName}: MapProps): JSX.Element {
   const mapRef = useRef(null);
   const map = useMap(mapRef, city);
+
+  useEffect(() => {
+    if (map) {
+      map.setView([
+        city.location.latitude,
+        city.location.longitude,
+      ],
+      city.location.zoom);
+    }
+  }, [city, map]);
 
   useEffect(() => {
     if (map) {
